# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: Campaign Operating Workspace API
...
servers:
  - url: /api
...
paths:
...
  /development/groups:
    get:
      operationId: getDevelopmentGroups
      parameters:
        - {$ref: "#/components/parameters/PageLimit"}
        - {$ref: "#/components/parameters/PageOffset"}
      responses:
        "200":
          description: Development teams, groups and unverified profile references; no memberships.
          headers:
            X-Pagination: {$ref: "#/components/headers/PopulationPage"}
          content:
            application/json:
              schema: {type: object, additionalProperties: true}
    post:
      operationId: createDevelopmentGroup
      requestBody:
        required: true
        content:
          application/json:
            schema: {$ref: "#/components/schemas/DevelopmentGroupInput"}
      responses:
        "201":
          description: Created unverified development organization unit.
          content:
            application/json:
              schema: {type: object, additionalProperties: true}
  /development/ownership/{type}/{id}:
    parameters:
      - {name: type, in: path, required: true, schema: {type: string, enum: [campaigns, activities]}}
      - {name: id, in: path, required: true, schema: {type: string, format: uuid}}
    get:
      operationId: getDevelopmentOwnership
      responses:
        "200":
          description: Stable ownership associations with unverified attribution.
          content:
            application/json:
              schema: {type: object, additionalProperties: true}
    put:
      operationId: updateDevelopmentOwnership
      requestBody:
        required: true
        content:
          application/json:
            schema: {$ref: "#/components/schemas/DevelopmentOwnershipInput"}
      responses:
        "200":
          description: Planning association changed; creator remains separate and unverified.
          content:
            application/json:
              schema: {type: object, additionalProperties: true}
  /development/calendar:
    get:
      operationId: getDevelopmentCalendar
      parameters:
        - {name: groupId, in: query, schema: {type: string, format: uuid}}
        - {name: campaignId, in: query, schema: {type: string, format: uuid}}
        - {name: after, in: query, schema: {type: string, format: uuid}}
        - {name: limit, in: query, schema: {type: integer, minimum: 1, maximum: 200, default: 100}}
      responses:
        "200":
          description: Stable activity-ID keyset page, filtered before counting. Live population; refresh from the first page to include newly inserted IDs preceding the cursor.
          content:
            application/json:
              schema:
                type: object
                required: [entries, total, returned, hasMore, nextCursor]
                properties:
                  entries: {type: array, items: {type: object, additionalProperties: true}}
                  total: {type: integer}
                  returned: {type: integer}
                  hasMore: {type: boolean}
                  nextCursor: {type: [string, "null"]}
                  unverified: {type: boolean}
                  authoritative: {type: boolean}
  /development/simulations:
    post:
      operationId: simulateDevelopmentWorkflow
      requestBody:
        required: true
        content:
          application/json:
            schema: {$ref: "#/components/schemas/DevelopmentSimulationInput"}
      responses:
        "200":
          description: Unverified dry-run; never approval, evidence or operational readiness.
          content:
            application/json:
              schema: {type: object, additionalProperties: true}
...
components:
  headers:
    PopulationPage:
      description: 'JSON object with total (authorized scoped records), returned (this response), hasMore and nextOffset (null at end). Stable ID ordered live pages; restart on population changes. For composite objects each top-level collection receives the same limit/offset independently, totals/returned sum the collections. Generated client consumes all pages, rejects partial failures, and deduplicates by ID.'
      schema: {type: string}
  parameters:
    PageLimit:
      name: limit
      in: query
      schema: {type: integer, minimum: 1, maximum: 500, default: 100}
    PageOffset:
      name: offset
      in: query
      schema: {type: integer, minimum: 0, default: 0}
...
  schemas:
    DevelopmentGroupInput:
      type: object
      additionalProperties: false
      required: [name, kind]
      properties:
        name: {type: string, minLength: 1, maxLength: 120}
        kind: {type: string, enum: [team, group]}
        parentId: {type: string, format: uuid}
        accountableOwnerId: {type: string, format: uuid}
    DevelopmentOwnershipInput:
      type: object
      additionalProperties: false
      required: [groupId, accountableOwnerId, rowVersion]
      properties:
        rowVersion: {type: integer, minimum: 1}
        groupId: {type: string, format: uuid}
        accountableOwnerId: {type: string, format: uuid}
    DevelopmentSimulationInput:
      type: object
      additionalProperties: false
      properties:
        campaignId: {type: string, format: uuid}
        activityId: {type: string, format: uuid}
        label: {type: string, maxLength: 120}
...