# lib/api-zod/src/generated/types/developmentOwnershipInput.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Api
 * Campaign Operating Workspace API
 * OpenAPI spec version: 0.1.0
 */

export interface DevelopmentOwnershipInput {
  /** @minimum 1 */
  rowVersion: number;
  groupId: string;
  accountableOwnerId: string;
}